feat(contract_manager): implement executeGovernanceInstruction for the SVM receiver
The receiver's governance messages are ExecutePostedVaa: post the VAA to
the core bridge its config names, then have the remote executor replay the
instructions it carries, signed by the PDA the config names as governance
authority. The claim record makes a re-run a no-op.

This moves executeThroughRemoteExecutor out of executor.ts and onto the
contract, which drops the core bridge lookup with it: the bridge to post to
is the one the receiver itself verifies against, which is also the owner the
remote executor requires of the posted VAA account.
